### 2019-05-16 #300 @noamross
Thank you for this inquiry, @MKLau!  I have a question: Can the package work without `provR`, as described in the README? If the two are in a tightly linked workflow, it may make more sense to have both reviewed or do them together.
