### 2019-05-17 #300 @MKLau
Thanks @noamross! No, but it does require some kind of provenance input. I have updated the package to use prospective provenance, which doesn't require code execution to conduct analyses. I would say that this is now the primary workhorse for the package. Sorry, I'm still working on the docs and haven't pushed the feature into a release yet, so it's not yet described. It depends heavily on the CodeDepends package by Duncan Lang. It's on CRAN, but I'm not sure if it's ROpenSci reviewed.
