### 2019-05-21 #300 @noamross
Hello @MKLau.  After some discussion, we've decided that this package is in-scope under the reproducibility category for converting analysis scripts to reproducible workflows.  I note that your package will require a vignette and to report [testing coverage](https://ropensci.github.io/dev_guide/ci.html#coverage) before being sent for review.
