### 2019-07-10 #300 @MKLau
@noamross

Hi Noam, just finished revisions to the package (including the addition of testing coverage and a vignette). Not sure what the next step is given your pre-review and my paused submission to JOSS.

Could you point me in the best direction?
