### 2019-07-24 #300 @noamross
My apologies for the delayed reply, @MKLau.  I see you opened another issue as a pre-submission inquiry, but you can go ahead and do a full submission for review. @annakrystalli is acting editor-in-chief while I'm traveling and will assign it out.
